+++ kubernetes-model-generator/openshift-model/cmd/generate/generate.go
@@ -33,6 +33,7 @@ import (
   userapi "github.com/openshift/api/user/v1"
   authenticationapi "k8s.io/api/authentication/v1"
   quotaapi "github.com/openshift/api/quota/v1"
+  helmapiv1beta1 "github.com/openshift/api/helm/v1beta1"
   rbac "k8s.io/api/rbac/v1" // depends
   "k8s.io/apimachinery/pkg/api/resource"
   metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -89,6 +90,10 @@ type Schema struct {
   RouteList                                routeapi.RouteList
   Template                                 templateapi.Template
   TemplateList                             templateapi.TemplateList
+  TemplateInstance                         templateapi.TemplateInstance
+  TemplateInstanceList                     templateapi.TemplateInstanceList
+  BrokerTemplateInstance                   templateapi.BrokerTemplateInstance
+  BrokerTemplateInstanceList               templateapi.BrokerTemplateInstanceList
   TagEvent                                 imageapi.TagEvent
   OAuthClient                              oauthapi.OAuthClient
   OAuthAccessToken                         oauthapi.OAuthAccessToken
@@ -125,6 +130,7 @@ type Schema struct {
   GroupList                                userapi.GroupList
   Identity                                 userapi.Identity
   IdentityList                             userapi.IdentityList
+  UserIdentityMapping                      userapi.UserIdentityMapping
   Config                                   configapi.Config
   RootPaths                                metav1.RootPaths
   Project                                  projectapi.Project
@@ -139,6 +145,8 @@ type Schema struct {
   ClusterNetworkList                       networkapi.ClusterNetworkList
   EgressNetworkPolicy                      networkapi.EgressNetworkPolicy
   EgressNetworkPolicyList                  networkapi.EgressNetworkPolicyList
+  HostSubnet                               networkapi.HostSubnet
+  HostSubnetList                           networkapi.HostSubnetList
   DNSZone                                  openshiftconfigapi.DNSZone
   ClusterVersion                           openshiftconfigapi.ClusterVersion
   ClusterVersionList                       openshiftconfigapi.ClusterVersionList
@@ -175,6 +183,8 @@ type Schema struct {
   AppliedClusterResourceQuotaList          quotaapi.AppliedClusterResourceQuotaList
   ClusterResourceQuota                     quotaapi.ClusterResourceQuota
   ClusterResourceQuotaList                 quotaapi.ClusterResourceQuotaList
+  HelmChartRepository                      helmapiv1beta1.HelmChartRepository
+  HelmChartRepositoryList                  helmapiv1beta1.HelmChartRepositoryList
 }
 
 func main() {
@@ -205,6 +215,7 @@ func main() {
     {"github.com/openshift/api/network/v1", "", "io.fabric8.openshift.api.model", "os_network_", true},
     {"github.com/openshift/api/config/v1", "", "io.fabric8.openshift.api.model", "os_config_v1_", true},
     {"github.com/openshift/api/quota/v1", "", "io.fabric8.openshift.api.model", "os_quota_", true},
+    {"github.com/openshift/api/helm/v1beta1", "", "io.fabric8.openshift.api.model", "os_helm_v1beta1_", true},
 
   }